Added new service for node creation
